pyproject.toml
[project]
name = "nonebot-plugin-summary-group"
version = "0.7.0"
authors = [{ name = "StillMisty", email = "1772797842@qq.com" }]
description = "基于 Nonebot2，使用 AI 分析群聊记录，生成讨论内容的总结。"
readme = "README.md"
license.file = "LICENSE"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Topic :: Communications :: Chat",
]
keywords = ["nonebot2", "nonebot", "plugin", "summary", "AI", "group-chat"]
dependencies = [
    "httpx>=0.27",
    "nonebot-adapter-onebot>=2.3",
    "nonebot2>=2.3",
    "nonebot_plugin_alconna>=0.53",
    "nonebot_plugin_localstore>=0.7",
    "nonebot_plugin_apscheduler>=0.5",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.package-data]
"nonebot_plugin_summary_group" = ["assert/*"]

[project.urls]
Homepage = "https://github.com/StillMisty/nonebot_plugin_summary_group"
